requirements.txt
alabaster==0.7.13
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
dash==2.9.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-leaflet==0.1.23
dash-table==5.0.0
exceptiongroup==1.1.1
Flask==2.2.3
geobuf==1.1.1
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
MouseInfo==0.1.3
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
plotly==5.13.1
protobuf==4.22.1
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.17
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.4
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rubicon-objc==0.4.5
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.3
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
validate-email==1.3
webdriver-manager==3.8.5
Werkzeug==2.2.3
wsproto==1.2.0
xgboost==1.7.4
zipp==3.15.0
